feature: migrate the API fetch-resource from core-services to dop
