+++ cmd/core-services/main.go
@@ -20,7 +20,6 @@ import (
 
 	// providers
 	_ "github.com/erda-project/erda-infra/providers/grpcclient"
-	_ "github.com/erda-project/erda-proto-go/cmp/dashboard/client"
 	_ "github.com/erda-project/erda-proto-go/core/pipeline/cms/client"
 	_ "github.com/erda-project/erda/modules/core-services"
 	_ "github.com/erda-project/erda/modules/core-services/services/notify/channel"
