+++ conf/dop/i18n/resource-trans.yaml
