+++ modules/core-services/services/org/org.go
@@ -16,10 +16,8 @@
 package org
 
 import (
-	"context"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/go-redis/redis"
@@ -27,7 +25,6 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/erda-project/erda-infra/providers/i18n"
-	dashboardPb "github.com/erda-project/erda-proto-go/cmp/dashboard/pb"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/modules/core-services/conf"
@@ -37,7 +34,6 @@ import (
 	"github.com/erda-project/erda/modules/core-services/types"
 	"github.com/erda-project/erda/pkg/crypto/uuid"
 	"github.com/erda-project/erda/pkg/numeral"
-	calcu "github.com/erda-project/erda/pkg/resourcecalculator"
 	"github.com/erda-project/erda/pkg/strutil"
 	"github.com/erda-project/erda/pkg/ucauth"
 )
@@ -49,8 +45,6 @@ type Org struct {
 	bdl      *bundle.Bundle
 	redisCli *redis.Client
 	trans    i18n.Translator
-
-	clusterResourceClient dashboardPb.ClusterResourceServer
 }
 
 // Option 定义 Org 对象的配置选项
@@ -90,13 +84,6 @@ func WithBundle(bdl *bundle.Bundle) Option {
 func WithRedisClient(cli *redis.Client) Option {
 	return func(o *Org) {
 		o.redisCli = cli
-	}
-}
-
-// WithClusterResourceClient set the gRPC client of CMP cluster resource
-func WithClusterResourceClient(cli dashboardPb.ClusterResourceServer) Option {
-	return func(o *Org) {
-		o.clusterResourceClient = cli
 	}
 }
 
@@ -703,211 +690,6 @@ func (o *Org) FetchOrgResources(orgID uint64) (*apistructs.OrgResourceInfo, erro
 		AvailableCpu: numeral.Round(totalCpu-usedCpu, 2),
 		AvailableMem: numeral.Round(totalMem-usedMem, 2),
 	}, nil
-}
-
-func (o *Org) FetchOrgClusterResource(ctx context.Context, orgID uint64) (*apistructs.OrgClustersResourcesInfo, error) {
-	langCodes, _ := ctx.Value("lang_codes").(i18n.LanguageCodes)
-
-	clusters, err := o.GetOrgClusterRelationsByOrg(orgID)
-	if err != nil {
-		return nil, err
-	}
-
-	var getClustersResourcesRequest dashboardPb.GetClustersResourcesRequest
-	for _, cluster := range clusters {
-		getClustersResourcesRequest.ClusterNames = append(getClustersResourcesRequest.ClusterNames, cluster.ClusterName)
-	}
-	// cmp gRPC 接口查询给定集群所有集群的资源和标签情况
-	resources, err := o.clusterResourceClient.GetClustersResources(ctx, &getClustersResourcesRequest)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to GetClusterResources, clusters: %v", getClustersResourcesRequest.GetClusterNames())
-	}
-
-	// 初始化所有集群的总资源 【】
-	var (
-		// 集群 nodes 数量
-		clusterNodes = make(map[string]int)
-		// 集群里资源计算器
-		calculators     = make(map[string]*calcu.Calculator)
-		requestResource = make(map[string]*calcu.Calculator)
-	)
-	countClustersNodes(clusterNodes, resources.List)
-	initClusterAllocatable(calculators, resources.List)
-	calculateRequest(requestResource, resources.List)
-
-	// 查出所有项目的 quota 记录
-	var projectsQuota []*model.ProjectQuota
-	if err = o.db.Find(&projectsQuota).Error; err != nil {
-		return nil, errors.Wrap(err, "failed to Find all project quota")
-	}
-	// 遍历所有项目和环境, 扣减对应集群的资源, 以计算出集群剩余资源
-	deductionQuota(calculators, projectsQuota)
-
-	var resourceInfo apistructs.OrgClustersResourcesInfo
-	for _, workspace := range calcu.Workspaces {
-		workspaceStr := calcu.WorkspaceString(workspace)
-
-		for _, clusterName := range getClustersResourcesRequest.ClusterNames {
-			calculator, ok := calculators[clusterName]
-			if !ok {
-				continue
-			}
-			resource := &apistructs.ClusterResources{
-				ClusterName:    clusterName,
-				Workspace:      workspaceStr,
-				CPUAllocatable: calcu.MillcoreToCore(calculator.AllocatableCPU(workspace), 3),
-				CPUAvailable:   calcu.MillcoreToCore(calculator.QuotableCPUForWorkspace(workspace), 3),
-				CPUQuotaRate:   0,
-				CPURequest:     0,
-				MemAllocatable: calcu.ByteToGibibyte(calculator.AllocatableMem(workspace), 3),
-				MemAvailable:   calcu.ByteToGibibyte(calculator.QuotableMemForWorkspace(workspace), 3),
-				MemQuotaRate:   0,
-				MemRequest:     0,
-				Nodes:          clusterNodes[clusterName],
-				Tips:           "",
-				CPUTookUp:      calcu.MillcoreToCore(calculator.AlreadyTookUpCPU(workspace), 3),
-				MemTookUp:      calcu.ByteToGibibyte(calculator.AlreadyTookUpMem(workspace), 3),
-			}
-			if c, ok := requestResource[clusterName]; ok {
-				resource.CPURequest = calcu.MillcoreToCore(c.AllocatableCPU(workspace), 3)
-				resource.MemRequest = calcu.ByteToGibibyte(c.AllocatableMem(workspace), 3)
-			}
-			if resource.CPUAllocatable > 0 {
-				resource.CPUQuotaRate = 1 - resource.CPUAvailable/resource.CPUAllocatable
-			}
-			if resource.MemAllocatable > 0 {
-				resource.MemQuotaRate = 1 - resource.MemAvailable/resource.MemAllocatable
-			}
-			o.makeTips(langCodes, resource, calculator, workspace)
-
-			resourceInfo.ClusterList = append(resourceInfo.ClusterList, resource)
-			resourceInfo.TotalCPU += resource.CPUAllocatable
-			resourceInfo.TotalMem += resource.MemAllocatable
-			resourceInfo.AvailableCPU += resource.CPUAvailable
-			resourceInfo.AvailableMem += resource.MemAvailable
-		}
-	}
-
-	return &resourceInfo, nil
-}
-
-func countClustersNodes(result map[string]int, list []*dashboardPb.ClusterResourceDetail) {
-	if result == nil {
-		return
-	}
-	for _, cluster := range list {
-		if !cluster.GetSuccess() {
-			logrus.WithField("cluster_name", cluster.GetClusterName()).WithField("err", cluster.GetErr()).
-				Warnln("the cluster is not valid now")
-			continue
-		}
-		result[cluster.GetClusterName()] = len(cluster.GetHosts())
-	}
-}
-
-func initClusterAllocatable(result map[string]*calcu.Calculator, list []*dashboardPb.ClusterResourceDetail) {
-	if result == nil {
-		return
-	}
-	for _, cluster := range list {
-		if !cluster.GetSuccess() {
-			logrus.WithField("cluster_name", cluster.GetClusterName()).WithField("err", cluster.GetErr()).
-				Warnln("the cluster is not valid now")
-			continue
-		}
-
-		// 累计此 host 上的 allocatable 资源
-		calculator, ok := result[cluster.GetClusterName()]
-		if !ok {
-			calculator = calcu.New(cluster.GetClusterName())
-		}
-		for _, host := range cluster.Hosts {
-			workspaces := extractWorkspacesFromLabels(host.GetLabels())
-			calculator.AddValue(host.GetCpuAllocatable(), host.GetMemAllocatable(), workspaces...)
-		}
-
-		result[cluster.GetClusterName()] = calculator
-	}
-}
-
-// 遍历所有项目和环境, 扣减对应集群的资源, 以计算出集群剩余资源
-func deductionQuota(clusters map[string]*calcu.Calculator, quotaRecords []*model.ProjectQuota) {
-	for _, workspace := range calcu.Workspaces {
-		workspaceStr := calcu.WorkspaceString(workspace)
-		for _, project := range quotaRecords {
-			if available, ok := clusters[project.GetClusterName(workspaceStr)]; ok {
-				available.DeductionQuota(workspace, project.GetCPUQuota(workspaceStr), project.GetMemQuota(workspaceStr))
-			}
-		}
-	}
-}
-
-func calculateRequest(result map[string]*calcu.Calculator, list []*dashboardPb.ClusterResourceDetail) {
-	if result == nil {
-		return
-	}
-	for _, cluster := range list {
-		if !cluster.GetSuccess() {
-			logrus.WithField("cluster_name", cluster.GetClusterName()).WithField("err", cluster.GetErr()).
-				Warnln("the cluster is not valid now")
-			continue
-		}
-
-		calculator, ok := result[cluster.GetClusterName()]
-		if !ok {
-			calculator = calcu.New(cluster.GetClusterName())
-		}
-		for _, host := range cluster.Hosts {
-			workspaces := extractWorkspacesFromLabels(host.GetLabels())
-			calculator.AddValue(host.GetCpuRequest(), host.GetMemRequest(), workspaces...)
-		}
-
-		result[cluster.GetClusterName()] = calculator
-	}
-}
-
-func (o *Org) makeTips(langCodes i18n.LanguageCodes, resource *apistructs.ClusterResources, calculator *calcu.Calculator,
-	workspace calcu.Workspace) {
-	if resource.CPUAllocatable == 0 && resource.MemAllocatable == 0 {
-		resource.Tips = o.trans.Text(langCodes, "NoResourceForTheWorkspace")
-		if resource.Nodes == 0 {
-			resource.Tips = o.trans.Text(langCodes, "NoNodesInTheCluster")
-		}
-		return
-	}
-
-	workspaceText := o.trans.Text(langCodes, strings.ToUpper(calcu.WorkspaceString(workspace)))
-	switch quotableCPU, quotableMem := calculator.QuotableCPUForWorkspace(workspace), calculator.QuotableMemForWorkspace(workspace); {
-	case quotableCPU == 0 || quotableMem == 0:
-		resource.Tips = fmt.Sprintf(o.trans.Text(langCodes, "ResourceSqueeze"), workspaceText, workspaceText)
-	case quotableCPU == 0:
-		resource.Tips = fmt.Sprintf(o.trans.Text(langCodes, "CPUResourceSqueeze"), workspaceText, workspaceText)
-	case quotableMem == 0:
-		resource.Tips = fmt.Sprintf(o.trans.Text(langCodes, "MemResourceSqueeze"), workspaceText, workspaceText)
-	}
-}
-
-func extractWorkspacesFromLabels(labels []string) []calcu.Workspace {
-	var (
-		m = make(map[calcu.Workspace]bool)
-		w []calcu.Workspace
-	)
-	for _, label := range labels {
-		switch strings.ToLower(label) {
-		case "dice/workspace-prod=true":
-			m[calcu.Prod] = true
-		case "dice/workspace-staging=true":
-			m[calcu.Staging] = true
-		case "dice/workspace-test=true":
-			m[calcu.Test] = true
-		case "dice/workspace-dev=true":
-			m[calcu.Dev] = true
-		}
-	}
-	for k := range m {
-		w = append(w, k)
-	}
-	return w
 }
 
 // ListAllOrgClusterRelation 获取所有企业对应集群关系
